+++ src/install/updates/update_v0.8.0_v0.x.x.php
@@ -10,7 +10,14 @@ require_once(dirname(__FILE__) . "/../../inc/Util.class.php");
 require_once(dirname(__FILE__) . "/../../inc/utils/AccessUtils.class.php");
 require_once(dirname(__FILE__) . "/../../inc/defines/config.php");
 
+echo "NOTICE: After this update the Peppers for Encryption.class.php and CSRF.class.php are stored in the new config file. So if you didn't merge them you would have to put the old pepper values into inc/conf.php to make the log in working again. Read more on this on the specific release information.\n";
+
 echo "Apply updates...\n";
+
+echo "Moving db config... ";
+rename(dirname(__FILE__)."/../../inc/db.php", dirname(__FILE__)."/../../inc/conf.php");
+file_put_contents(dirname(__FILE__)."/../../inc/conf.php", "\n".'$PEPPER = ["__PEPPER1__","__PEPPER2__","__PEPPER3__","__CSRF__"];'."\n", FILE_APPEND);
+echo "OK\n";
 
 echo "Check agent binaries... ";
 Util::checkAgentVersion("python", "0.2.0");